Update/Fix/Add for drawable,Java Class and XML Class
- Add 2 new icons

- Update XML files

- Update Manifest for some Activities

- Fix switch page function for Product Sub Module
