+++ app/src/main/AndroidManifest.xml
@@ -25,19 +25,16 @@
         </activity>
         <activity
             android:name=".RegisterActivity"
-            android:exported="true" />
+            android:exported="false" />
+        <activity
+            android:name=".ProductMenuSubModuleActivity"
+            android:exported="false"/>
+        <activity
+            android:name=".ProductListActivity"
+            android:exported="false"/>
         <activity
             android:name=".MainActivity"
-            android:exported="true">
-
-            <!--
-            intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter
-            -->
-        </activity>
+            android:exported="false"/>
     </application>
 
 </manifest>
