+++ app/src/main/res/layout/activity_product_management.xml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@drawable/dot_background"
     tools:context=".ProductManagementActivity">
 
     <LinearLayout
@@ -34,7 +35,7 @@
                 android:layout_height="45dp"
                 android:layout_margin="16dp"
                 android:layout_weight="0.333"
-                android:background="@drawable/grey_background"
+                android:background="@drawable/bottom_background"
                 android:elevation="5dp"
                 android:orientation="vertical">
 
@@ -66,15 +67,8 @@
         android:layout_height="wrap_content"
         android:layout_gravity="bottom|end"
         android:layout_margin="16dp"
-        android:layout_marginTop="4dp"
-        android:src="@drawable/add_circle"
+        android:src="@drawable/add"
         app:backgroundTint="@color/teal_200"
-        app:elevation="6dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.005"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:tint="@android:color/white" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+        app:tint="@android:color/white"
+        app:elevation="6dp" />
+</FrameLayout>
